# Improves the logic to import objects in helpers