--- a/packages/babel-preset-env/test/fixtures/dynamic-import/modules-amd/output.js
+++ b/packages/babel-preset-env/test/fixtures/dynamic-import/modules-amd/output.js
@@ -1,7 +1,7 @@
 define(["require"], function (_require) {
-  function _createRequireWildcardCache(obj) { if (typeof global._REQUIRE_WILD_CACHE === 'undefined') { global._REQUIRE_WILD_CACHE = typeof WeakMap === 'function' ? new WeakMap() : null; } var cache = global._REQUIRE_WILD_CACHE; return { has: function has() { return cache ? cache.has(obj) : false; }, set: function set(newObj) { cache && cache.set(obj, newObj); }, get: function get() { return cache ? cache.get(obj) : null; } }; }
+  function _getRequireWildcardCache(obj) { if (!global || typeof WeakMap !== 'function') { return null; } if (!global._REQUIRE_WILD_CACHE) { global._REQUIRE_WILD_CACHE = new WeakMap(); } return global._REQUIRE_WILD_CACHE; }
 
-  function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } var cache = _createRequireWildcardCache(obj); if (cache.has()) { return cache.get(); } var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) { var desc = Object.defineProperty && Object.getOwnPropertyDescriptor ? Object.getOwnPropertyDescriptor(obj, key) : {}; if (desc.get || desc.set) { Object.defineProperty(newObj, key, desc); } else { newObj[key] = obj[key]; } } } } newObj["default"] = obj; cache.set(newObj); return newObj; }
+  function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } var cache = _getRequireWildcardCache(obj); if (cache && cache.has(obj)) { return cache.get(obj); } var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) { var desc = Object.defineProperty && Object.getOwnPropertyDescriptor ? Object.getOwnPropertyDescriptor(obj, key) : {}; if (desc.get || desc.set) { Object.defineProperty(newObj, key, desc); } else { newObj[key] = obj[key]; } } } } newObj["default"] = obj; cache && cache.set(obj, newObj); return newObj; }
 
   new Promise(function (_resolve, _reject) {
     return _require(["foo"], function (imported) {