… packages/babel-preset-env/test/fixtures/dynamic-import/modules-cjs/output.js