| Q                        | A <!--(Can use an emoji 👍) -->
| ------------------------ | ---
| Fixed Issues?            | No
| Patch: Bug Fix?          | No
| Major: Breaking Change?  | No
| Minor: New Feature?      | No
| Tests Added + Pass?      | Yes
| Documentation PR Link    | N/A
| Any Dependency Changes?  | No
| License                  | MIT

---------

In the case of `react-native`,

```js
import React, { PureComponent } from 'react';
```

Compiled as below.

```js
var _react = _interopRequireWildcard(_$$_REQUIRE(_dependencyMap[7], "react"));
```

The below logic(looping) is performed whenever `_interopRequireWildcard` with `react` is called.

```js
var newObj = {};

if (obj != null) {
  for (var key in obj) {
    if (Object.prototype.hasOwnProperty.call(obj, key)) {
      var desc = Object.defineProperty && Object.getOwnPropertyDescriptor ? Object.getOwnPropertyDescriptor(obj, key) : {};

      if (desc.get || desc.set) {
        Object.defineProperty(newObj, key, desc);
      } else {
        newObj[key] = obj[key];
      }
    }
  }
}
```

Generally, We have a lot of components that imports `react` or `react-native`. I don't think it's good to check the same object repeatedly on every call. so, I thought it would be better to put cache logic in here.

I counted the number of how many loops are executed to start my production-level App. It may feel like, but it feels like my app is a little faster. :)

|Version|Count
| ------------------------ | ---
|Current | 11,032|
|New (Cache) | 128|

I don't fully understand `Babel` architecture. Please let me know if I have any mistakes.

I hope this helps.