@nicolo-ribaudo 2019-07-04
Note that this optimization would only work when using `@babel/plugin-transform-runtime`, otherwise the map is re-created in every file.