@ifsnow 2019-07-05
@nicolo-ribaudo I've modified it to work well in the case you mentioned. Please let me know if there's anything else I need to consider.