@ifsnow 2019-07-16
@nicolo-ribaudo @Andarist @zloirock Thank you all for the feedback.

I realized that there might be no `global` object. https://github.com/webpack/webpack/blob/master/buildin/global.js#L16

In this case, I modified the cache to not work. I'm sure `cache` works well in modern development environments.