@ifsnow 2019-08-08
@nicolo-ribaudo @jridgewell @Andarist @zloirock @ljharb @cpojer Do you have plans to merge this PR?

I'm sure this cache feature makes React Native apps run faster. The following table shows the time spent running js bundle at startup.

| Device + Engine              | Current (ms) | Cache (ms) | Improving |
| -----------------------------|-------------:| ----------:|----------:|
| Android 6.0 (32bit) + JSC    | 3,182        | **2,990**      |6%🔻 |
| Android 7.0 (64bit) + JSC    | 1,118        | **1,089**      | 2.6%🔻 |
| Android 6.0 (32bit) + Hermes | 1,230        | **964**      |21.7%🔻 |
| Android 7.0 (64bit) + Hermes | 403          | **367**       |9%🔻 |

The following cases are more effective.
- Use many components
- Use low performance devices
- Use Hermes engine

I think we can improve the performance of many React Native apps without any effort. Please review it positively.