@ifsnow 2019-08-28
@nicolo-ribaudo Do you have any plans to release the new version? I'm wondering when this fix will be available :)